Voice_Assistant/requirements.txt
azure-cognitiveservices-speech
openai
python-dotenv
pyaudio
django